Error_Repair_MySQL_Connection

+++ src/main/java/com/example/auctionservicesaplication/model/Category.java
@@ -22,7 +22,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_id")
-    private BigDecimal id;
+    private Long id;
 
     @OneToMany(mappedBy = "category")
     private List<Auction> auctions;
