+++ src/main/java/com/example/auctionservicesaplication/model/TransactionRating.java
@@ -19,7 +19,7 @@ public class TransactionRating {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "transaction_rating_id")
-    private BigDecimal id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
